% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{dir_info_par}
\alias{dir_info_par}
\title{Returns a dataframe of file information for all files in the selected directory recursively}
\usage{
dir_info_par(path)
}
\description{
Returns a dataframe of file information for all files in the selected directory recursively
}
